requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
google-generativeai==0.8.3
python-dotenv==1.0.1
sqlite3

